+++ src/components/search-button/searchButton.css
@@ -1,5 +1,4 @@
 .search__button {
-  margin-right: 20px;
   border-radius: 5px;
 
   color: white;
